main-site/api/requirementscopy.txt
Flask==3.0.2
Flask_Login==0.6.3
oauthlib==3.2.2
python-dotenv==1.0.1
Requests==2.31.0
gunicorn==20.0.4
azure-storage-blob
pyodbc
azure-data-tables
azure-identity
openai==1.12.0
python-dotenv==1.0.1
SQLAlchemy==2.0.27
Flask-SQLAlchemy==3.1.1
psycopg2
bcrypt
